refactor(app): split mask export helpers
